refactor(benchmarks): extract shared oracle/compare helpers into lib
